refactor: move OperationStatus to pkg/execution
Move OperationStatus enum from pkg/types to pkg/execution/operation_status.go
as it's specific to execution context. Update all references across:
- pkg/types/execution_context.go
- pkg/handlerpipeline (execute.go, integration.go)
- Test files

Used 'exec' alias in handlerpipeline to avoid naming conflict with local variable.

Part of #704
